agents/pyproject.toml
[project]
name = "contract-swarm-agents"
version = "0.1.0"
description = "ContractSwarm AI agent backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "aiosqlite>=0.20.0",
    "pdfplumber>=0.11.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
thenvoi = [
    "thenvoi-sdk[claude_sdk]",
    "claude-agent-sdk",
]

[tool.uv.sources]
thenvoi-sdk = { git = "https://github.com/thenvoi/thenvoi-sdk-python.git" }
